# backend/swagger.yaml
openapi: 3.0.1
info:
  title: FluxPilot API
  description: "REST API платформы FluxPilot для управления workflow, их версиями\
    , снапшотами, триггерами, запусками и A/B-аналитикой."
  version: 1.0.0
servers:
- url: /v1
  description: Основной API сервер
...
paths:
...
  /workflows/{workflowId}/runs:
    get:
      tags:
      - Runs
      operationId: listRuns
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowRun"
    post:
      tags:
      - Runs
      operationId: runWorkflow
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: startNodeId
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonNode"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowRun"
  /workflows/{workflowId}/nodes/{nodeId}/debug-run:
    post:
      tags:
      - Debug
      summary: Отладочный запуск одной ноды
      description: Исполняет одну выбранную ноду с произвольным входом без прогона
        графа; результат возвращается синхронно и фиксируется в истории как отладочный
        запуск.
      operationId: debugRunNode
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: nodeId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebugNodeRunRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DebugNodeRunResult"
  /workflows/{workflowId}/debug-sessions:
    post:
      tags:
      - Debug
      operationId: start
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebugStartRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DebugSessionDto"
  /webhook/{token}:
    post:
      tags:
      - Triggers
      operationId: webhook
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonNode"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebhookAccepted"
  /debug-sessions/{sessionId}/step:
    post:
      tags:
      - Debug
      operationId: step
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebugStepRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DebugSessionDto"
  /debug-sessions/{sessionId}/run-to-end:
    post:
      tags:
      - Debug
      operationId: runToEnd
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DebugSessionDto"
  /workflows/{workflowId}/triggers/{triggerId}:
    patch:
      tags:
      - Triggers
      operationId: update
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: triggerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerUpdate"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Trigger"
  /workflows/{workflowId}/triggers:
    get:
      tags:
      - Triggers
      operationId: list_2
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trigger"
...
  /workflow-runs/{runId}:
    get:
      tags:
      - Runs
      operationId: getRun
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowRun"
  /workflow-runs/{runId}/result:
    get:
      tags:
      - Runs
      operationId: getRunResult
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowRunResult"
  /node-runs/{nodeRunId}:
    get:
      tags:
      - Runs
      operationId: getNodeRun
      parameters:
      - name: nodeRunId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NodeRun"
  /debug-sessions/{sessionId}:
    get:
      tags:
      - Debug
      operationId: get_2
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DebugSessionDto"
    delete:
      tags:
      - Debug
      operationId: close
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    JsonNode:
      type: object
...
    NodeRun:
      required:
      - id
      - nodeId
      - status
      - workflowRunId
      type: object
      properties:
        id:
          type: string
        workflowRunId:
          type: string
        nodeId:
          type: string
        status:
          type: string
        startedAt:
          type: string
        finishedAt:
          type: string
        input:
          $ref: "#/components/schemas/JsonNode"
        output:
          $ref: "#/components/schemas/JsonNode"
        errorMessage:
          type: string
    WorkflowRun:
      required:
      - id
      - isDebug
      - nodes
      - status
      - workflowId
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        status:
          type: string
        startedAt:
          type: string
        finishedAt:
          type: string
        durationMs:
          type: integer
          format: int64
        input:
          $ref: "#/components/schemas/JsonNode"
        output:
          $ref: "#/components/schemas/JsonNode"
        startNodeId:
          type: string
        isDebug:
          type: boolean
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeRun"
    DebugNodeRunRequest:
      type: object
      properties:
        input:
          $ref: "#/components/schemas/JsonNode"
    DebugNodeRunResult:
      required:
      - nodeId
      - runId
      - status
      - workflowId
      type: object
      properties:
        runId:
          type: string
        workflowId:
          type: string
        nodeId:
          type: string
        status:
          type: string
        input:
          $ref: "#/components/schemas/JsonNode"
        output:
          $ref: "#/components/schemas/JsonNode"
        errorMessage:
          type: string
    DebugStartRequest:
      type: object
      properties:
        input:
          $ref: "#/components/schemas/JsonNode"
        startNodeId:
          type: string
    DebugFailedNode:
      required:
      - message
      - nodeId
      type: object
      properties:
        nodeId:
          type: string
        message:
          type: string
    DebugSessionDto:
      required:
      - completed
      - createdAt
      - failed
      - outputs
      - ready
      - readyInputs
      - sessionId
      - skipped
      - status
      - updatedAt
      - versionId
      - workflowId
      type: object
      properties:
        sessionId:
          type: string
        workflowId:
          type: string
        versionId:
          type: string
        status:
          type: string
        input:
          $ref: "#/components/schemas/JsonNode"
        outputs:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/JsonNode"
        completed:
          type: array
          items:
            type: string
        skipped:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            $ref: "#/components/schemas/DebugFailedNode"
        ready:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        readyInputs:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/JsonNode"
    WebhookAccepted:
      required:
      - pollUrl
      - run
      type: object
      properties:
        run:
          $ref: "#/components/schemas/WorkflowRun"
        pollUrl:
          type: string
    DebugStepRequest:
      type: object
      properties:
        nodeId:
          type: string
    TriggerUpdate:
      required:
      - enabled
      type: object
      properties:
        enabled:
          type: boolean
    Trigger:
      required:
      - enabled
      - id
      - nodeId
      - type
      - workflowId
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        nodeId:
          type: string
        type:
          type: string
        config:
          $ref: "#/components/schemas/JsonNode"
        token:
          type: string
        enabled:
          type: boolean
...
    WorkflowRunResult:
      required:
      - id
      - status
      - workflowId
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        status:
          type: string
        startedAt:
          type: string
        finishedAt:
          type: string
        durationMs:
          type: integer
          format: int64
        output:
          $ref: "#/components/schemas/JsonNode"